Rename methods in Catalog C API
